replace strncpy with strlcpy + truncation checks
The strncpy function doesn't guarantee NUL termination and manual NUL
termination was missing for two of these calls. Truncation should also
be treated as an error in all but rare cases where it's desired. The
strlcpy function returns the size of the source string, making it easy
to check for this.

Change-Id: If9b4fc12cf1582c33f5e81faddaf12346f9265ad